meta: convert attributeToSql and attributesToSql to TypeScript
Moves both methods out of every dialect's `query-generator.js` and into its
`query-generator-typescript.internal.ts`, and renames them from `attributeToSQL`
/ `attributesToSQL` to `attributeToSql` / `attributesToSql`.

Closes the last item of #15533.

The core declaration is now a typed contract every dialect implements, rather
than a `ColumnsDescription` signature that described none of them. The new
`AttributeToSqlColumn` interface documents what the dialects actually read, and
is deliberately wider than `NormalizedAttributeOptions` because these methods
are also called with `describeTable` output, where the type is a raw SQL string.

`isBinaryStringType` joins `attributeTypeToDataTypeId` in the core data type
utilities, replacing the `attribute.type.options?.binary` reach-through.

No SQL changes. The pinned suites produce byte-identical output for all nine
dialects. Typing the tests did surface two things they were getting wrong:
a `deferrable` passed as a raw string rather than the `Deferrable` enum, and a
leftover pre-rename option name.

Three no-op arguments went away because the compiler proved they could not do
anything: `AttributeToSqlOptions` and `EscapeOptions` share no properties, so
mssql and snowflake were handing `escape` an object it could never read, and
several `toSql({ dialect })` calls passed an argument no implementation accepts.
