+++ packages/core/src/abstract-dialect/query-generator.d.ts
@@ -16,10 +16,13 @@ import type {
   AbstractQueryGeneratorTypeScript,
   ParameterOptions,
 } from './query-generator-typescript.js';
-import type { AttributeToSqlOptions } from './query-generator.internal-types.js';
+import type {
+  AttributesToSqlColumns,
+  AttributeToSqlInput,
+  AttributeToSqlOptions,
+} from './query-generator.internal-types.js';
 import type { BoundQuery, TableOrModel } from './query-generator.types.js';
 import type { TableName } from './query-interface.js';
-import type { ColumnsDescription } from './query-interface.types.js';
 import type { WhereOptions } from './where-sql-builder-types.js';
 
 type SelectOptions<M extends Model> = FindOptions<M> & {
@@ -128,8 +131,10 @@ export class AbstractQueryGenerator<
     options?: CreateTableQueryOptions,
   ): string;
 
-  attributesToSQL(
-    attributes: ColumnsDescription,
+  attributeToSql(column: AttributeToSqlInput, options?: AttributeToSqlOptions): string;
+
+  attributesToSql(
+    columns: AttributesToSqlColumns,
     options?: AttributeToSqlOptions,
   ): Record<string, string>;
 }
