+++ packages/core/src/abstract-dialect/query-interface.js
@@ -95,7 +95,7 @@ export class AbstractQueryInterface extends AbstractQueryInterfaceTypeScript {
       tableName.schema = modelTable?.schema || options.schema;
     }
 
-    attributes = this.queryGenerator.attributesToSQL(attributes, {
+    attributes = this.queryGenerator.attributesToSql(attributes, {
       tableOrModel: tableName,
       context: 'createTable',
       withoutForeignKeyConstraints: options.withoutForeignKeyConstraints,
@@ -206,7 +206,7 @@ export class AbstractQueryInterface extends AbstractQueryInterfaceTypeScript {
     options ||= {};
 
     const columns = { [attributeName]: this.normalizeAttribute(dataTypeOrOptions) };
-    const query = this.queryGenerator.attributesToSQL(columns, {
+    const query = this.queryGenerator.attributesToSql(columns, {
       context: 'changeColumn',
       tableOrModel: tableName,
     });
@@ -266,7 +266,7 @@ export class AbstractQueryInterface extends AbstractQueryInterfaceTypeScript {
     const sql = this.queryGenerator.renameColumnQuery(
       tableName,
       attrNameBefore,
-      this.queryGenerator.attributesToSQL(_options),
+      this.queryGenerator.attributesToSql(_options),
     );
 
     return await this.sequelize.queryRaw(sql, options);
