+++ packages/db2/src/query-generator-typescript.internal.ts
@@ -9,17 +9,31 @@ import type {
   TableOrModel,
   TruncateTableQueryOptions,
 } from '@sequelize/core';
-import { AbstractQueryGenerator, Op } from '@sequelize/core';
+import { AbstractQueryGenerator, DataTypes, Op } from '@sequelize/core';
+import { attributeTypeToSql } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
+import type {
+  ENUM,
+  NormalizedDataType,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types.js';
+import type { EscapeOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
 import {
   DROP_SCHEMA_QUERY_SUPPORTABLE_OPTIONS,
   REMOVE_INDEX_QUERY_SUPPORTABLE_OPTIONS,
   RENAME_TABLE_QUERY_SUPPORTABLE_OPTIONS,
   TRUNCATE_TABLE_QUERY_SUPPORTABLE_OPTIONS,
 } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
+import type {
+  AttributesToSqlColumns,
+  AttributeToSqlColumn,
+  AttributeToSqlInput,
+  AttributeToSqlOptions,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator.internal-types.js';
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { joinSQLFragments } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/join-sql-fragments.js';
 import { EMPTY_SET } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
+import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { generateIndexName } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
+import isPlainObject from 'lodash/isPlainObject';
 import { randomBytes } from 'node:crypto';
 import type { Db2Dialect } from './dialect.js';
 import { Db2QueryGeneratorInternal } from './query-generator.internal.js';
@@ -258,5 +272,126 @@ export class Db2QueryGeneratorTypeScript extends AbstractQueryGenerator {
 
   getRandomFloatFunctionCall(): string {
     return 'RAND()';
+  }
+
+  attributeToSql(column: AttributeToSqlInput, options?: AttributeToSqlOptions): string {
+    const attribute: AttributeToSqlColumn = isPlainObject(column)
+      ? (column as AttributeToSqlColumn)
+      : { type: column as NormalizedDataType };
+
+    let template: string;
+
+    if (attribute.type instanceof DataTypes.ENUM) {
+      // enums are a special case
+      const enumType = attribute.type as ENUM<string>;
+
+      template = enumType.toSql();
+      template += ` CHECK (${this.quoteIdentifier(attribute.field!)} IN(${enumType.options.values
+        .map(value => {
+          return this.escape(value);
+        })
+        .join(', ')}))`;
+    } else {
+      template = attributeTypeToSql(attribute.type);
+    }
+
+    if (options && options.context === 'changeColumn' && attribute.type) {
+      template = `DATA TYPE ${template}`;
+    } else if (attribute.allowNull === false || attribute.primaryKey === true) {
+      template += ' NOT NULL';
+    }
+
+    if (attribute.autoIncrement) {
+      let initialValue: number | string = 1;
+      if (attribute.initialAutoIncrement) {
+        initialValue = attribute.initialAutoIncrement;
+      }
+
+      template += ` GENERATED BY DEFAULT AS IDENTITY(START WITH ${initialValue}, INCREMENT BY 1)`;
+    }
+
+    if (defaultValueSchemable(attribute.defaultValue, this.dialect)) {
+      // `replacements` is not part of AttributeToSqlOptions, but the JS version read it
+      template += ` DEFAULT ${this.escape(attribute.defaultValue, { replacements: (options as EscapeOptions | undefined)?.replacements, type: attribute.type })}`;
+    }
+
+    if (
+      attribute.unique === true &&
+      (options?.context !== 'changeColumn' || this.dialect.supports.alterColumn.unique)
+    ) {
+      template += ' UNIQUE';
+    }
+
+    if (attribute.primaryKey) {
+      template += ' PRIMARY KEY';
+    }
+
+    if (!options?.withoutForeignKeyConstraints && attribute.references) {
+      if (options?.context === 'addColumn' && options.tableOrModel && attribute.field) {
+        const fkName = this.quoteIdentifier(
+          `${this.extractTableDetails(options.tableOrModel).tableName}_${attribute.field}_fidx`,
+        );
+
+        template += ` ADD CONSTRAINT ${fkName} FOREIGN KEY (${this.quoteIdentifier(attribute.field)})`;
+      }
+
+      template += ` REFERENCES ${this.quoteTable(attribute.references.table)}`;
+
+      if (attribute.references.key) {
+        template += ` (${this.quoteIdentifier(attribute.references.key)})`;
+      } else {
+        template += ` (${this.quoteIdentifier('id')})`;
+      }
+
+      if (attribute.onDelete) {
+        template += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
+      }
+
+      if (attribute.onUpdate && attribute.onUpdate.toUpperCase() !== 'CASCADE') {
+        // Db2 do not support CASCADE option for ON UPDATE clause.
+        template += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
+      }
+    }
+
+    if (attribute.comment && typeof attribute.comment === 'string') {
+      template += ` COMMENT ${attribute.comment}`;
+    }
+
+    return template;
+  }
+
+  attributesToSql(
+    columns: AttributesToSqlColumns,
+    options?: AttributeToSqlOptions,
+  ): Record<string, string> {
+    const result: Record<string, string> = Object.create(null);
+    const existingConstraints: string[] = [];
+
+    for (const key of Object.keys(columns)) {
+      const rawColumn = columns[key];
+      const attribute: AttributeToSqlColumn = isPlainObject(rawColumn)
+        ? { ...(rawColumn as AttributeToSqlColumn) }
+        : { type: rawColumn as NormalizedDataType };
+      const columnName = attribute.field || attribute.columnName || key;
+
+      attribute.field = columnName;
+
+      if (attribute.references) {
+        if (existingConstraints.includes(this.quoteTable(attribute.references.table))) {
+          // db2 rejects more than one cascading constraint to the same table
+          attribute.onDelete = '';
+          attribute.onUpdate = '';
+        } else if (attribute.unique === true) {
+          attribute.onDelete = '';
+          attribute.onUpdate = '';
+        } else {
+          existingConstraints.push(this.quoteTable(attribute.references.table));
+        }
+      }
+
+      result[columnName] = this.attributeToSql(attribute, options);
+    }
+
+    return result;
   }
 }
