+++ packages/db2/src/query-generator.js
@@ -1,10 +1,7 @@
 'use strict';
 
 import { DataTypes, Op, ParameterStyle } from '@sequelize/core';
-import {
-  attributeTypeToSql,
-  normalizeDataType,
-} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
+import { normalizeDataType } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
 import {
   ADD_COLUMN_QUERY_SUPPORTABLE_OPTIONS,
   CREATE_TABLE_QUERY_SUPPORTABLE_OPTIONS,
@@ -12,7 +9,6 @@ import {
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { removeNullishValuesFromHash } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/format.js';
 import { EMPTY_SET } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
-import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { createBindParamGenerator } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/sql.js';
 import { removeTrailingSemicolon } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
 import { pojo } from '@sequelize/utils';
@@ -21,7 +17,6 @@ import each from 'lodash/each';
 import forEach from 'lodash/forEach';
 import forOwn from 'lodash/forOwn';
 import includes from 'lodash/includes';
-import isPlainObject from 'lodash/isPlainObject';
 import isString from 'lodash/isString';
 import startsWith from 'lodash/startsWith';
 import template from 'lodash/template';
@@ -162,7 +157,7 @@ export class Db2QueryGenerator extends Db2QueryGeneratorTypeScript {
 
     dataType = {
       ...dataType,
-      // TODO: attributeToSQL SHOULD be using attributes in addColumnQuery
+      // TODO: attributeToSql SHOULD be using attributes in addColumnQuery
       //       but instead we need to pass the key along as the field here
       field: key,
       type: normalizeDataType(dataType.type, this.dialect),
@@ -174,7 +169,7 @@ export class Db2QueryGenerator extends Db2QueryGeneratorTypeScript {
       this._templateSettings,
     )({
       key: this.quoteIdentifier(key),
-      definition: this.attributeToSQL(dataType, {
+      definition: this.attributeToSql(dataType, {
         context: 'addColumn',
         tableOrModel: table,
       }),
@@ -543,121 +538,6 @@ export class Db2QueryGenerator extends Db2QueryGeneratorTypeScript {
     }
 
     return super.addIndexQuery(tableName, attributes, options, rawTablename);
-  }
-
-  attributeToSQL(attribute, options) {
-    if (!isPlainObject(attribute)) {
-      attribute = {
-        type: attribute,
-      };
-    }
-
-    let template;
-
-    if (attribute.type instanceof DataTypes.ENUM) {
-      // enums are a special case
-      template = attribute.type.toSql({ dialect: this.dialect });
-      template += ` CHECK (${this.quoteIdentifier(attribute.field)} IN(${attribute.type.options.values
-        .map(value => {
-          return this.escape(value, undefined, { replacements: options?.replacements });
-        })
-        .join(', ')}))`;
-    } else {
-      template = attributeTypeToSql(attribute.type);
-    }
-
-    if (options && options.context === 'changeColumn' && attribute.type) {
-      template = `DATA TYPE ${template}`;
-    } else if (attribute.allowNull === false || attribute.primaryKey === true) {
-      template += ' NOT NULL';
-    }
-
-    if (attribute.autoIncrement) {
-      let initialValue = 1;
-      if (attribute.initialAutoIncrement) {
-        initialValue = attribute.initialAutoIncrement;
-      }
-
-      template += ` GENERATED BY DEFAULT AS IDENTITY(START WITH ${initialValue}, INCREMENT BY 1)`;
-    }
-
-    if (defaultValueSchemable(attribute.defaultValue, this.dialect)) {
-      template += ` DEFAULT ${this.escape(attribute.defaultValue, { replacements: options?.replacements, type: attribute.type })}`;
-    }
-
-    if (
-      attribute.unique === true &&
-      (options?.context !== 'changeColumn' || this.dialect.supports.alterColumn.unique)
-    ) {
-      template += ' UNIQUE';
-    }
-
-    if (attribute.primaryKey) {
-      template += ' PRIMARY KEY';
-    }
-
-    if (!options?.withoutForeignKeyConstraints && attribute.references) {
-      if (options?.context === 'addColumn' && attribute.field) {
-        const fkName = this.quoteIdentifier(
-          `${this.extractTableDetails(options.tableOrModel).tableName}_${attribute.field}_fidx`,
-        );
-
-        template += ` ADD CONSTRAINT ${fkName} FOREIGN KEY (${this.quoteIdentifier(attribute.field)})`;
-      }
-
-      template += ` REFERENCES ${this.quoteTable(attribute.references.table)}`;
-
-      if (attribute.references.key) {
-        template += ` (${this.quoteIdentifier(attribute.references.key)})`;
-      } else {
-        template += ` (${this.quoteIdentifier('id')})`;
-      }
-
-      if (attribute.onDelete) {
-        template += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
-      }
-
-      if (attribute.onUpdate && attribute.onUpdate.toUpperCase() !== 'CASCADE') {
-        // Db2 do not support CASCADE option for ON UPDATE clause.
-        template += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
-      }
-    }
-
-    if (attribute.comment && typeof attribute.comment === 'string') {
-      template += ` COMMENT ${attribute.comment}`;
-    }
-
-    return template;
-  }
-
-  attributesToSQL(attributes, options) {
-    const result = {};
-    const existingConstraints = [];
-
-    for (const key of Object.keys(attributes)) {
-      const rawAttribute = attributes[key];
-      const attribute = isPlainObject(rawAttribute) ? { ...rawAttribute } : { type: rawAttribute };
-      const columnName = attribute.field || attribute.columnName || key;
-
-      attribute.field = columnName;
-
-      if (attribute.references) {
-        if (existingConstraints.includes(this.quoteTable(attribute.references.table))) {
-          // db2 rejects more than one cascading constraint to the same table
-          attribute.onDelete = '';
-          attribute.onUpdate = '';
-        } else if (attribute.unique === true) {
-          attribute.onDelete = '';
-          attribute.onUpdate = '';
-        } else {
-          existingConstraints.push(this.quoteTable(attribute.references.table));
-        }
-      }
-
-      result[columnName] = this.attributeToSQL(attribute, options);
-    }
-
-    return result;
   }
 
   createTrigger() {
