+++ packages/ibmi/src/query-generator-typescript.internal.ts
@@ -7,18 +7,37 @@ import type {
   TableOrModel,
   TruncateTableQueryOptions,
 } from '@sequelize/core';
-import { AbstractQueryGenerator } from '@sequelize/core';
+import { AbstractQueryGenerator, DataTypes } from '@sequelize/core';
+import {
+  attributeTypeToDataTypeId,
+  attributeTypeToSql,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
+import type {
+  ENUM,
+  NormalizedDataType,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types.js';
 import {
   REMOVE_INDEX_QUERY_SUPPORTABLE_OPTIONS,
   RENAME_TABLE_QUERY_SUPPORTABLE_OPTIONS,
   TRUNCATE_TABLE_QUERY_SUPPORTABLE_OPTIONS,
 } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
+import type {
+  AttributesToSqlColumns,
+  AttributeToSqlColumn,
+  AttributeToSqlInput,
+  AttributeToSqlOptions,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator.internal-types.js';
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { joinSQLFragments } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/join-sql-fragments.js';
 import { EMPTY_SET } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
+import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { generateIndexName } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
+import { pojo } from '@sequelize/utils';
+import isPlainObject from 'lodash/isPlainObject';
 import type { IBMiDialect } from './dialect.js';
 import { IBMiQueryGeneratorInternal } from './query-generator.internal.js';
+
+const typeWithoutDefault = new Set(['BLOB']);
 
 const REMOVE_INDEX_QUERY_SUPPORTED_OPTIONS = new Set<keyof RemoveIndexQueryOptions>(['ifExists']);
 
@@ -233,5 +252,125 @@ export class IBMiQueryGeneratorTypeScript extends AbstractQueryGenerator {
 
   getRandomFloatFunctionCall(): string {
     return 'RAND()';
+  }
+
+  attributeToSql(column: AttributeToSqlInput, options?: AttributeToSqlOptions): string {
+    const attribute: AttributeToSqlColumn = isPlainObject(column)
+      ? (column as AttributeToSqlColumn)
+      : { type: column as NormalizedDataType };
+
+    const attributeString = attribute.type.toString();
+    let template = attributeString;
+
+    if (attribute.type instanceof DataTypes.ENUM) {
+      // enums are a special case
+      const enumType = attribute.type as ENUM<string>;
+
+      template = enumType.toSql();
+      if (options && options.context) {
+        template += options.context === 'changeColumn' ? ' ADD' : '';
+      }
+
+      template += ` CHECK (${this.quoteIdentifier(attribute.field!)} IN(${enumType.options.values
+        .map(value => {
+          return this.escape(value);
+        })
+        .join(', ')}))`;
+    } else {
+      template = attributeTypeToSql(attribute.type);
+    }
+
+    // in the changeColumn context the nullability is emitted as its own ALTER COLUMN clause
+    if (attribute.allowNull === false && options?.context !== 'changeColumn') {
+      template += ' NOT NULL';
+    }
+
+    if (attribute.autoIncrement) {
+      template += ' GENERATED BY DEFAULT AS IDENTITY (START WITH 1, INCREMENT BY 1)';
+    }
+
+    // BLOB cannot have a default value
+    if (
+      !typeWithoutDefault.has(attributeTypeToDataTypeId(attribute.type)) &&
+      defaultValueSchemable(attribute.defaultValue, this.dialect)
+    ) {
+      if (attribute.defaultValue === true) {
+        attribute.defaultValue = 1;
+      } else if (attribute.defaultValue === false) {
+        attribute.defaultValue = 0;
+      }
+
+      template += ` DEFAULT ${this.escape(attribute.defaultValue)}`;
+    }
+
+    if (attribute.unique === true && !attribute.primaryKey) {
+      template += ' UNIQUE';
+    }
+
+    if (attribute.primaryKey) {
+      template += ' PRIMARY KEY';
+    }
+
+    // Db2 for i comments are a mess
+    // if (attribute.comment) {
+    //   template += ` ${options.context === 'changeColumn' ? 'ADD ' : ''}COMMENT ${this.escape(attribute.comment)}`;
+    // }
+
+    if (attribute.first) {
+      template += ' FIRST';
+    }
+
+    if (attribute.after) {
+      template += ` AFTER ${this.quoteIdentifier(attribute.after)}`;
+    }
+
+    if (!options?.withoutForeignKeyConstraints && attribute.references) {
+      if (options?.context === 'addColumn' && options.tableOrModel && attribute.field) {
+        const fkName = this.quoteIdentifier(
+          `${this.extractTableDetails(options.tableOrModel).tableName}_${attribute.field}_foreign_idx`,
+        );
+
+        template += ` ADD CONSTRAINT ${fkName} FOREIGN KEY (${this.quoteIdentifier(attribute.field)})`;
+      }
+
+      template += ` REFERENCES ${this.quoteTable(attribute.references.table)}`;
+
+      if (attribute.references.key) {
+        template += ` (${this.quoteIdentifier(attribute.references.key)})`;
+      } else {
+        template += ` (${this.quoteIdentifier('id')})`;
+      }
+
+      if (attribute.onDelete) {
+        template += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
+      }
+
+      if (attribute.onUpdate && attribute.onUpdate.toUpperCase() !== 'CASCADE') {
+        template += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
+      }
+    }
+
+    return template;
+  }
+
+  attributesToSql(
+    columns: AttributesToSqlColumns,
+    options?: AttributeToSqlOptions,
+  ): Record<string, string> {
+    const result: Record<string, string> = pojo();
+
+    for (const key of Object.keys(columns)) {
+      const rawColumn = columns[key];
+      const attribute: AttributeToSqlColumn = isPlainObject(rawColumn)
+        ? { ...(rawColumn as AttributeToSqlColumn) }
+        : { type: rawColumn as NormalizedDataType };
+      const columnName = attribute.field || attribute.columnName || key;
+
+      attribute.field = columnName;
+
+      result[columnName] = this.attributeToSql(attribute, options);
+    }
+
+    return result;
   }
 }
