+++ packages/ibmi/src/query-generator.js
@@ -1,11 +1,6 @@
 'use strict';
 
-import { DataTypes } from '@sequelize/core';
-import {
-  attributeTypeToDataTypeId,
-  attributeTypeToSql,
-  normalizeDataType,
-} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
+import { normalizeDataType } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
 import {
   ADD_COLUMN_QUERY_SUPPORTABLE_OPTIONS,
   CREATE_TABLE_QUERY_SUPPORTABLE_OPTIONS,
@@ -14,18 +9,14 @@ import { BaseSqlExpression } from '@sequelize/core/_non-semver-use-at-your-own-r
 import { conformIndex } from '@sequelize/core/_non-semver-use-at-your-own-risk_/model-internals.js';
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { EMPTY_SET } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
-import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import {
   nameIndex,
   removeTrailingSemicolon,
 } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
 import { pojo } from '@sequelize/utils';
 import each from 'lodash/each';
-import isPlainObject from 'lodash/isPlainObject';
 import util from 'node:util';
 import { IBMiQueryGeneratorTypeScript } from './query-generator-typescript.internal.js';
-
-const typeWithoutDefault = new Set(['BLOB']);
 
 const CREATE_TABLE_QUERY_SUPPORTED_OPTIONS = new Set(['uniqueKeys']);
 
@@ -124,13 +115,13 @@ export class IBMiQueryGenerator extends IBMiQueryGeneratorTypeScript {
 
     dataType = {
       ...dataType,
-      // TODO: attributeToSQL SHOULD be using attributes in addColumnQuery
+      // TODO: attributeToSql SHOULD be using attributes in addColumnQuery
       //       but instead we need to pass the key along as the field here
       field: key,
       type: normalizeDataType(dataType.type, this.dialect),
     };
 
-    const definition = this.attributeToSQL(dataType, {
+    const definition = this.attributeToSql(dataType, {
       context: 'addColumn',
       tableOrModel: table,
     });
@@ -353,123 +344,5 @@ export class IBMiQueryGenerator extends IBMiQueryGeneratorTypeScript {
     }
 
     return query;
-  }
-
-  attributeToSQL(attribute, options) {
-    if (!isPlainObject(attribute)) {
-      attribute = {
-        type: attribute,
-      };
-    }
-
-    const attributeString = attribute.type.toString({
-      escape: this.escape.bind(this),
-      dialect: this.dialect,
-    });
-    let template = attributeString;
-
-    if (attribute.type instanceof DataTypes.ENUM) {
-      // enums are a special case
-      template = attribute.type.toSql({ dialect: this.dialect });
-      if (options && options.context) {
-        template += options.context === 'changeColumn' ? ' ADD' : '';
-      }
-
-      template += ` CHECK (${this.quoteIdentifier(attribute.field)} IN(${attribute.type.options.values
-        .map(value => {
-          return this.escape(value);
-        })
-        .join(', ')}))`;
-    } else {
-      template = attributeTypeToSql(attribute.type);
-    }
-
-    // in the changeColumn context the nullability is emitted as its own ALTER COLUMN clause
-    if (attribute.allowNull === false && options?.context !== 'changeColumn') {
-      template += ' NOT NULL';
-    }
-
-    if (attribute.autoIncrement) {
-      template += ' GENERATED BY DEFAULT AS IDENTITY (START WITH 1, INCREMENT BY 1)';
-    }
-
-    // BLOB cannot have a default value
-    if (
-      !typeWithoutDefault.has(attributeTypeToDataTypeId(attribute.type)) &&
-      defaultValueSchemable(attribute.defaultValue, this.dialect)
-    ) {
-      if (attribute.defaultValue === true) {
-        attribute.defaultValue = 1;
-      } else if (attribute.defaultValue === false) {
-        attribute.defaultValue = 0;
-      }
-
-      template += ` DEFAULT ${this.escape(attribute.defaultValue)}`;
-    }
-
-    if (attribute.unique === true && !attribute.primaryKey) {
-      template += ' UNIQUE';
-    }
-
-    if (attribute.primaryKey) {
-      template += ' PRIMARY KEY';
-    }
-
-    // Db2 for i comments are a mess
-    // if (attribute.comment) {
-    //   template += ` ${options.context === 'changeColumn' ? 'ADD ' : ''}COMMENT ${this.escape(attribute.comment)}`;
-    // }
-
-    if (attribute.first) {
-      template += ' FIRST';
-    }
-
-    if (attribute.after) {
-      template += ` AFTER ${this.quoteIdentifier(attribute.after)}`;
-    }
-
-    if (!options?.withoutForeignKeyConstraints && attribute.references) {
-      if (options?.context === 'addColumn' && attribute.field) {
-        const fkName = this.quoteIdentifier(
-          `${this.extractTableDetails(options.tableOrModel).tableName}_${attribute.field}_foreign_idx`,
-        );
-
-        template += ` ADD CONSTRAINT ${fkName} FOREIGN KEY (${this.quoteIdentifier(attribute.field)})`;
-      }
-
-      template += ` REFERENCES ${this.quoteTable(attribute.references.table)}`;
-
-      if (attribute.references.key) {
-        template += ` (${this.quoteIdentifier(attribute.references.key)})`;
-      } else {
-        template += ` (${this.quoteIdentifier('id')})`;
-      }
-
-      if (attribute.onDelete) {
-        template += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
-      }
-
-      if (attribute.onUpdate && attribute.onUpdate.toUpperCase() !== 'CASCADE') {
-        template += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
-      }
-    }
-
-    return template;
-  }
-
-  attributesToSQL(attributes, options) {
-    const result = pojo();
-
-    for (const key of Object.keys(attributes)) {
-      const rawAttribute = attributes[key];
-      const attribute = isPlainObject(rawAttribute) ? { ...rawAttribute } : { type: rawAttribute };
-      const columnName = attribute.field || attribute.columnName || key;
-
-      attribute.field = columnName;
-
-      result[columnName] = this.attributeToSQL(attribute, options);
-    }
-
-    return result;
   }
 }
