+++ packages/mssql/src/query-generator-typescript.internal.ts
@@ -12,18 +12,29 @@ import type {
   TableOrModel,
   TruncateTableQueryOptions,
 } from '@sequelize/core';
-import { AbstractQueryGenerator } from '@sequelize/core';
+import { AbstractQueryGenerator, DataTypes } from '@sequelize/core';
+import { attributeTypeToSql } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
+import type { NormalizedDataType } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types.js';
 import type { EscapeOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
 import {
   CREATE_DATABASE_QUERY_SUPPORTABLE_OPTIONS,
   REMOVE_INDEX_QUERY_SUPPORTABLE_OPTIONS,
   TRUNCATE_TABLE_QUERY_SUPPORTABLE_OPTIONS,
 } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
+import type {
+  AttributesToSqlColumns,
+  AttributeToSqlColumn,
+  AttributeToSqlInput,
+  AttributeToSqlOptions,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator.internal-types.js';
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { joinSQLFragments } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/join-sql-fragments.js';
 import { buildJsonPath } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/json.js';
 import { EMPTY_SET } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
+import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { generateIndexName } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
+import { pojo } from '@sequelize/utils';
+import isPlainObject from 'lodash/isPlainObject';
 import { randomBytes } from 'node:crypto';
 import type { MsSqlDialect } from './dialect.js';
 import { MsSqlQueryGeneratorInternal } from './query-generator.internal.js';
@@ -348,5 +359,134 @@ SELECT REVERSE(SUBSTRING(@ms_ver, CHARINDEX('.', @ms_ver)+1, 20)) AS 'version'`;
 
   getRandomFloatFunctionCall(): string {
     return 'RAND()';
+  }
+
+  attributeToSql(column: AttributeToSqlInput, options?: AttributeToSqlOptions): string {
+    const attribute: AttributeToSqlColumn = isPlainObject(column)
+      ? (column as AttributeToSqlColumn)
+      : { type: column as NormalizedDataType };
+
+    // handle self-referential constraints
+    if (
+      attribute.references &&
+      attribute.Model &&
+      this.isSameTable(
+        (attribute.Model as { tableName: string }).tableName,
+        attribute.references.table,
+      )
+    ) {
+      this.sequelize.log(
+        'MSSQL does not support self-referential constraints, ' +
+          'we will remove it but we recommend restructuring your query',
+      );
+      attribute.onDelete = '';
+      attribute.onUpdate = '';
+    }
+
+    let template: string;
+
+    if (attribute.type instanceof DataTypes.ENUM) {
+      // enums are a special case
+      template = attribute.type.toSql();
+      template += ` CHECK (${this.quoteIdentifier(attribute.field!)} IN(${attribute.type.options.values
+        .map(value => {
+          // these options are not escape options, but they were passed through in the JS version
+          return this.escape(value);
+        })
+        .join(', ')}))`;
+
+      return template;
+    }
+
+    template = attributeTypeToSql(attribute.type);
+
+    if (attribute.allowNull === false) {
+      template += ' NOT NULL';
+    } else if (
+      !attribute.primaryKey &&
+      !defaultValueSchemable(attribute.defaultValue, this.dialect)
+    ) {
+      template += ' NULL';
+    }
+
+    if (attribute.autoIncrement) {
+      template += ' IDENTITY(1,1)';
+    }
+
+    if (defaultValueSchemable(attribute.defaultValue, this.dialect)) {
+      template += ` DEFAULT ${this.escape(attribute.defaultValue, { type: attribute.type })}`;
+    }
+
+    if (
+      attribute.unique === true &&
+      (options?.context !== 'changeColumn' || this.dialect.supports.alterColumn.unique)
+    ) {
+      template += ' UNIQUE';
+    }
+
+    if (attribute.primaryKey) {
+      template += ' PRIMARY KEY';
+    }
+
+    if (!options?.withoutForeignKeyConstraints && attribute.references) {
+      template += ` REFERENCES ${this.quoteTable(attribute.references.table)}`;
+
+      if (attribute.references.key) {
+        template += ` (${this.quoteIdentifier(attribute.references.key)})`;
+      } else {
+        template += ` (${this.quoteIdentifier('id')})`;
+      }
+
+      if (attribute.onDelete) {
+        template += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
+      }
+
+      if (attribute.onUpdate) {
+        template += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
+      }
+    }
+
+    if (attribute.comment && typeof attribute.comment === 'string') {
+      template += ` COMMENT ${attribute.comment}`;
+    }
+
+    return template;
+  }
+
+  attributesToSql(
+    columns: AttributesToSqlColumns,
+    options?: AttributeToSqlOptions,
+  ): Record<string, string> {
+    const result: Record<string, string> = pojo();
+    const existingConstraints: string[] = [];
+
+    for (const key of Object.keys(columns)) {
+      const rawColumn = columns[key];
+      const attribute: AttributeToSqlColumn = isPlainObject(rawColumn)
+        ? { ...(rawColumn as AttributeToSqlColumn) }
+        : { type: rawColumn as NormalizedDataType };
+      const columnName = attribute.field || attribute.columnName || key;
+
+      attribute.field = columnName;
+
+      if (attribute.references) {
+        if (existingConstraints.includes(this.quoteTable(attribute.references.table))) {
+          // mssql rejects more than one cascading constraint to the same table
+          attribute.onDelete = '';
+          attribute.onUpdate = '';
+        } else {
+          existingConstraints.push(this.quoteTable(attribute.references.table));
+
+          // NOTE: this really just disables cascading updates for all
+          //       definitions. Can be made more robust to support the
+          //       few cases where MSSQL actually supports them
+          attribute.onUpdate = '';
+        }
+      }
+
+      result[columnName] = this.attributeToSql(attribute, options);
+    }
+
+    return result;
   }
 }
