+++ packages/oracle/src/query-generator-typescript.internal.ts
@@ -11,7 +11,8 @@ import type {
   TableOrModel,
   TruncateTableQueryOptions,
 } from '@sequelize/core';
-import { AbstractQueryGenerator, IsolationLevel } from '@sequelize/core';
+import { AbstractQueryGenerator, DataTypes, IsolationLevel } from '@sequelize/core';
+import type { NormalizedDataType } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types.js';
 import {
   CREATE_SCHEMA_QUERY_SUPPORTABLE_OPTIONS,
   REMOVE_COLUMN_QUERY_SUPPORTABLE_OPTIONS,
@@ -19,6 +20,12 @@ import {
   RENAME_TABLE_QUERY_SUPPORTABLE_OPTIONS,
   TRUNCATE_TABLE_QUERY_SUPPORTABLE_OPTIONS,
 } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
+import type {
+  AttributesToSqlColumns,
+  AttributeToSqlColumn,
+  AttributeToSqlInput,
+  AttributeToSqlOptions,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator.internal-types.js';
 import type { TableNameWithSchema } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-interface.js';
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { joinSQLFragments } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/join-sql-fragments.js';
@@ -27,7 +34,9 @@ import {
   isModelStatic,
 } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/model-utils.js';
 import { EMPTY_SET } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
+import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { generateIndexName } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
+import isPlainObject from 'lodash/isPlainObject';
 import type { OracleDialect } from './dialect.js';
 import { OracleQueryGeneratorInternal } from './query-generator.internal.js';
 
@@ -409,5 +418,132 @@ export class OracleQueryGeneratorTypeScript extends AbstractQueryGenerator {
 
   getRandomFloatFunctionCall(): string {
     return 'DBMS_RANDOM.VALUE()';
+  }
+
+  attributeToSql(column: AttributeToSqlInput, options?: AttributeToSqlOptions): string {
+    const attribute: AttributeToSqlColumn = isPlainObject(column)
+      ? (column as AttributeToSqlColumn)
+      : { type: column as NormalizedDataType };
+
+    // handle self referential constraints
+    const model = (attribute as unknown as { Model?: { tableName?: string } }).Model;
+    if (
+      attribute.references &&
+      model &&
+      model.tableName === (attribute.references as unknown as { tableName?: string }).tableName
+    ) {
+      this.sequelize.log(
+        'Oracle does not support self referencial constraints, ' +
+          'we will remove it but we recommend restructuring your query',
+      );
+      attribute.onDelete = '';
+    }
+
+    let template: string;
+
+    if (attribute.type instanceof DataTypes.ENUM) {
+      // enums are a special case
+      template = attribute.type.toSql();
+      template += ` CHECK (${this.quoteIdentifier(attribute.field!)} IN(${attribute.type.options.values
+        .map(value => {
+          return this.escape(value);
+        })
+        .join(', ')}))`;
+
+      return template;
+    }
+
+    if (attribute.type instanceof DataTypes.JSON) {
+      template = attribute.type.toSql();
+      template += ` CHECK (${this.quoteIdentifier(attribute.field!)} IS JSON)`;
+
+      return template;
+    }
+
+    if (attribute.type instanceof DataTypes.BOOLEAN) {
+      template = attribute.type.toSql();
+      template += ` CHECK (${this.quoteIdentifier(attribute.field!)} IN('1', '0'))`;
+
+      return template;
+    }
+
+    if (attribute.autoIncrement) {
+      template = ' NUMBER(*,0) GENERATED BY DEFAULT ON NULL AS IDENTITY';
+    } else if (attribute.type && attribute.type === 'DOUBLE') {
+      // this branch only runs for the raw string 'DOUBLE', which has no toSql method
+      template = (attribute.type as unknown as { toSql(): string }).toSql();
+    } else if (attribute.type) {
+      // setting it to false because oracle doesn't support unsigned int so put a check to make it behave like unsigned int
+      let unsignedTemplate = '';
+      const typeOptions = (attribute.type as unknown as { options?: { unsigned?: boolean } })
+        .options;
+      if (typeOptions?.unsigned) {
+        typeOptions.unsigned = false;
+        unsignedTemplate += ` CHECK(${this.quoteIdentifier(attribute.field!)} >= 0)`;
+      }
+
+      template = attribute.type.toString();
+
+      if (attribute.type && defaultValueSchemable(attribute.defaultValue, this.dialect)) {
+        template += ` DEFAULT ${this.escape(attribute.defaultValue)}`;
+      }
+
+      if (!attribute.autoIncrement) {
+        // If autoincrement, not null is set automatically
+        if (attribute.allowNull === false) {
+          template += ' NOT NULL';
+        } else if (
+          !attribute.primaryKey &&
+          !defaultValueSchemable(attribute.defaultValue, this.dialect)
+        ) {
+          template += ' NULL';
+        }
+      }
+
+      template += unsignedTemplate;
+    } else {
+      template = '';
+    }
+
+    if (attribute.primaryKey) {
+      template += ' PRIMARY KEY';
+    }
+
+    if (!options?.withoutForeignKeyConstraints && attribute.references) {
+      template += ` REFERENCES ${this.quoteTable(attribute.references.table)}`;
+
+      if (attribute.references.key) {
+        template += ` (${this.quoteIdentifier(attribute.references.key)})`;
+      } else {
+        template += ` (${this.quoteIdentifier('id')})`;
+      }
+
+      if (attribute.onDelete && attribute.onDelete.toUpperCase() !== 'NO ACTION') {
+        template += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
+      }
+    }
+
+    return template;
+  }
+
+  attributesToSql(
+    columns: AttributesToSqlColumns,
+    options?: AttributeToSqlOptions,
+  ): Record<string, string> {
+    const result: Record<string, string> = Object.create(null);
+
+    for (const key of Object.keys(columns)) {
+      const rawColumn = columns[key];
+      const attribute: AttributeToSqlColumn = isPlainObject(rawColumn)
+        ? { ...(rawColumn as AttributeToSqlColumn) }
+        : { type: rawColumn as NormalizedDataType };
+      const columnName = attribute.field || attribute.columnName || key;
+
+      attribute.field = columnName;
+
+      result[columnName] = this.attributeToSql(attribute, options);
+    }
+
+    return result;
   }
 }
