+++ packages/oracle/src/query-generator.js
@@ -5,11 +5,9 @@
 import each from 'lodash/each';
 import forOwn from 'lodash/forOwn';
 import includes from 'lodash/includes';
-import isPlainObject from 'lodash/isPlainObject';
 import toPath from 'lodash/toPath';
 import oracledb from 'oracledb';
 
-import { DataTypes } from '@sequelize/core';
 import { normalizeDataType } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
 import {
   ADD_COLUMN_QUERY_SUPPORTABLE_OPTIONS,
@@ -23,7 +21,6 @@ import {
   EMPTY_SET,
   getObjectFromMap,
 } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
-import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { pojo } from '@sequelize/utils';
 import { OracleQueryGeneratorTypeScript } from './query-generator-typescript.internal';
 
@@ -440,7 +437,7 @@ export class OracleQueryGenerator extends OracleQueryGeneratorTypeScript {
 
     const attribute = joinSQLFragments([
       this.quoteIdentifier(key),
-      this.attributeToSQL(dataType, {
+      this.attributeToSql(dataType, {
         context: 'addColumn',
         tableOrModel: table,
       }),
@@ -844,126 +841,6 @@ export class OracleQueryGenerator extends OracleQueryGeneratorTypeScript {
     }
 
     return queryTmpl;
-  }
-
-  attributeToSQL(attribute, options) {
-    if (!isPlainObject(attribute)) {
-      attribute = {
-        type: attribute,
-      };
-    }
-
-    // handle self referential constraints
-    if (
-      attribute.references &&
-      attribute.Model &&
-      attribute.Model.tableName === attribute.references.tableName
-    ) {
-      this.sequelize.log(
-        'Oracle does not support self referencial constraints, ' +
-          'we will remove it but we recommend restructuring your query',
-      );
-      attribute.onDelete = '';
-    }
-
-    let template;
-
-    if (attribute.type instanceof DataTypes.ENUM) {
-      // enums are a special case
-      template = attribute.type.toSql({ dialect: this.dialect });
-      template += ` CHECK (${this.quoteIdentifier(attribute.field)} IN(${attribute.type.options.values
-        .map(value => {
-          return this.escape(value, undefined, {});
-        })
-        .join(', ')}))`;
-
-      return template;
-    }
-
-    if (attribute.type instanceof DataTypes.JSON) {
-      template = attribute.type.toSql();
-      template += ` CHECK (${this.quoteIdentifier(attribute.field)} IS JSON)`;
-
-      return template;
-    }
-
-    if (attribute.type instanceof DataTypes.BOOLEAN) {
-      template = attribute.type.toSql();
-      template += ` CHECK (${this.quoteIdentifier(attribute.field)} IN('1', '0'))`;
-
-      return template;
-    }
-
-    if (attribute.autoIncrement) {
-      template = ' NUMBER(*,0) GENERATED BY DEFAULT ON NULL AS IDENTITY';
-    } else if (attribute.type && attribute.type === 'DOUBLE') {
-      template = attribute.type.toSql();
-    } else if (attribute.type) {
-      // setting it to false because oracle doesn't support unsigned int so put a check to make it behave like unsigned int
-      let unsignedTemplate = '';
-      if (attribute.type?.options?.unsigned) {
-        attribute.type.options.unsigned = false;
-        unsignedTemplate += ` CHECK(${this.quoteIdentifier(attribute.field)} >= 0)`;
-      }
-
-      template = attribute.type.toString();
-
-      if (attribute.type && defaultValueSchemable(attribute.defaultValue, this.dialect)) {
-        template += ` DEFAULT ${this.escape(attribute.defaultValue)}`;
-      }
-
-      if (!attribute.autoIncrement) {
-        // If autoincrement, not null is set automatically
-        if (attribute.allowNull === false) {
-          template += ' NOT NULL';
-        } else if (
-          !attribute.primaryKey &&
-          !defaultValueSchemable(attribute.defaultValue, this.dialect)
-        ) {
-          template += ' NULL';
-        }
-      }
-
-      template += unsignedTemplate;
-    } else {
-      template = '';
-    }
-
-    if (attribute.primaryKey) {
-      template += ' PRIMARY KEY';
-    }
-
-    if (!options?.withoutForeignKeyConstraints && attribute.references) {
-      template += ` REFERENCES ${this.quoteTable(attribute.references.table)}`;
-
-      if (attribute.references.key) {
-        template += ` (${this.quoteIdentifier(attribute.references.key)})`;
-      } else {
-        template += ` (${this.quoteIdentifier('id')})`;
-      }
-
-      if (attribute.onDelete && attribute.onDelete.toUpperCase() !== 'NO ACTION') {
-        template += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
-      }
-    }
-
-    return template;
-  }
-
-  attributesToSQL(attributes, options) {
-    const result = {};
-
-    for (const key of Object.keys(attributes)) {
-      const rawAttribute = attributes[key];
-      const attribute = isPlainObject(rawAttribute) ? { ...rawAttribute } : { type: rawAttribute };
-      const columnName = attribute.field || attribute.columnName || key;
-
-      attribute.field = columnName;
-
-      result[columnName] = this.attributeToSQL(attribute, options);
-    }
-
-    return result;
   }
 
   createTrigger() {
