+++ packages/postgres/src/query-generator-typescript.internal.ts
@@ -10,12 +10,24 @@ import type {
   TableOrModel,
   TruncateTableQueryOptions,
 } from '@sequelize/core';
-import { AbstractQueryGenerator } from '@sequelize/core';
+import { AbstractQueryGenerator, DataTypes } from '@sequelize/core';
+import type {
+  ENUM,
+  NormalizedDataType,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types.js';
 import type { EscapeOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
 import { CREATE_DATABASE_QUERY_SUPPORTABLE_OPTIONS } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator-typescript.js';
+import type {
+  AttributesToSqlColumns,
+  AttributeToSqlColumn,
+  AttributeToSqlInput,
+  AttributeToSqlOptions,
+} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator.internal-types.js';
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { joinSQLFragments } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/join-sql-fragments.js';
+import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { generateIndexName } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
+import isPlainObject from 'lodash/isPlainObject';
 import semver from 'semver';
 import type { PostgresDialect } from './dialect.js';
 import { PostgresQueryGeneratorInternal } from './query-generator.internal.js';
@@ -290,5 +302,135 @@ export class PostgresQueryGeneratorTypeScript extends AbstractQueryGenerator {
 
   versionQuery() {
     return 'SHOW SERVER_VERSION';
+  }
+
+  attributeToSql(column: AttributeToSqlInput, options?: AttributeToSqlOptions): string {
+    const attribute: AttributeToSqlColumn = isPlainObject(column)
+      ? (column as AttributeToSqlColumn)
+      : { type: column as NormalizedDataType };
+
+    let type: NormalizedDataType | undefined;
+    const arraySubtype =
+      attribute.type instanceof DataTypes.ARRAY ? attribute.type.options.type : null;
+
+    if (
+      attribute.type instanceof DataTypes.ENUM ||
+      (attribute.type instanceof DataTypes.ARRAY && arraySubtype instanceof DataTypes.ENUM)
+    ) {
+      const enumType = (arraySubtype || attribute.type) as ENUM<string>;
+      const values = enumType.options.values;
+
+      if (Array.isArray(values) && values.length > 0) {
+        type = `ENUM(${values.map(value => this.escape(value)).join(', ')})`;
+
+        if (attribute.type instanceof DataTypes.ARRAY) {
+          type += '[]';
+        }
+      } else {
+        throw new Error("Values for ENUM haven't been defined.");
+      }
+    }
+
+    if (!type) {
+      type = attribute.type;
+    }
+
+    let sql = type.toString();
+
+    if (attribute.allowNull === false) {
+      sql += ' NOT NULL';
+    }
+
+    if (attribute.autoIncrement) {
+      if (attribute.autoIncrementIdentity) {
+        sql += ' GENERATED BY DEFAULT AS IDENTITY';
+      } else {
+        sql += ' SERIAL';
+      }
+    }
+
+    if (defaultValueSchemable(attribute.defaultValue, this.dialect)) {
+      sql += ` DEFAULT ${this.escape(attribute.defaultValue, { type: attribute.type })}`;
+    }
+
+    if (attribute.unique === true) {
+      sql += ' UNIQUE';
+    }
+
+    if (attribute.primaryKey) {
+      sql += ' PRIMARY KEY';
+    }
+
+    if (attribute.references) {
+      // an unqualified reference target lives in the same schema as the table that references it
+      const schema =
+        options?.tableOrModel &&
+        (!attribute.references.table || typeof attribute.references.table === 'string')
+          ? this.extractTableDetails(options.tableOrModel).schema
+          : undefined;
+
+      const referencesTable = this.extractTableDetails(attribute.references.table, {
+        schema,
+      } as { schema?: string });
+
+      let referencesKey;
+
+      if (!options?.withoutForeignKeyConstraints) {
+        if (attribute.references.key) {
+          referencesKey = this.quoteIdentifiers(attribute.references.key);
+        } else {
+          referencesKey = this.quoteIdentifier('id');
+        }
+
+        sql += ` REFERENCES ${this.quoteTable(referencesTable)} (${referencesKey})`;
+
+        if (attribute.onDelete) {
+          sql += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
+        }
+
+        if (attribute.onUpdate) {
+          sql += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
+        }
+
+        if (attribute.references.deferrable) {
+          sql += ` ${this.#internals.getDeferrableConstraintSnippet(attribute.references.deferrable)}`;
+        }
+      }
+    }
+
+    if (attribute.comment && typeof attribute.comment === 'string') {
+      if (options?.context === 'addColumn' || options?.context === 'changeColumn') {
+        const quotedAttr = this.quoteIdentifier(attribute.field!);
+        const escapedCommentText = this.escape(attribute.comment);
+        sql += `; COMMENT ON COLUMN ${this.quoteTable(options.tableOrModel!)}.${quotedAttr} IS ${escapedCommentText}`;
+      } else {
+        // for createTable event which does it's own parsing
+        // TODO: centralize creation of comment statements here
+        sql += ` COMMENT ${attribute.comment}`;
+      }
+    }
+
+    return sql;
+  }
+
+  attributesToSql(
+    columns: AttributesToSqlColumns,
+    options?: AttributeToSqlOptions,
+  ): Record<string, string> {
+    const result: Record<string, string> = Object.create(null);
+
+    for (const key of Object.keys(columns)) {
+      const rawColumn = columns[key];
+      const attribute: AttributeToSqlColumn = isPlainObject(rawColumn)
+        ? { ...(rawColumn as AttributeToSqlColumn) }
+        : { type: rawColumn as NormalizedDataType };
+      const columnName = attribute.field || attribute.columnName || key;
+
+      attribute.field = columnName;
+
+      result[columnName] = this.attributeToSql(attribute, options);
+    }
+
+    return result;
   }
 }
