+++ packages/postgres/src/query-generator.js
@@ -4,11 +4,9 @@ import { DataTypes } from '@sequelize/core';
 import { CREATE_TABLE_QUERY_SUPPORTABLE_OPTIONS } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/query-generator.js';
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { quoteIdentifier } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/dialect.js';
-import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { generateIndexName } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/string.js';
 import each from 'lodash/each';
 import isEmpty from 'lodash/isEmpty';
-import isPlainObject from 'lodash/isPlainObject';
 import map from 'lodash/map';
 import reduce from 'lodash/reduce';
 import { ENUM } from './_internal/data-types-overrides.js';
@@ -114,7 +112,7 @@ export class PostgresQueryGenerator extends PostgresQueryGeneratorTypeScript {
   addColumnQuery(table, key, attribute, options) {
     options ||= {};
 
-    const dbDataType = this.attributeToSQL(
+    const dbDataType = this.attributeToSql(
       { ...attribute, field: attribute.field || key },
       { context: 'addColumn', tableOrModel: table },
     );
@@ -214,131 +212,6 @@ export class PostgresQueryGenerator extends PostgresQueryGeneratorTypeScript {
     parameters ||= '';
 
     return `CREATE OR REPLACE FUNCTION pg_temp.${fnName}(${parameters}) ${returns} AS $func$ BEGIN ${body} END; $func$ LANGUAGE ${language}; SELECT * FROM pg_temp.${fnName}();`;
-  }
-
-  attributeToSQL(attribute, options) {
-    if (!isPlainObject(attribute)) {
-      attribute = {
-        type: attribute,
-      };
-    }
-
-    let type;
-    const arraySubtype =
-      attribute.type instanceof DataTypes.ARRAY ? attribute.type.options.type : null;
-
-    if (
-      attribute.type instanceof DataTypes.ENUM ||
-      (attribute.type instanceof DataTypes.ARRAY && arraySubtype instanceof DataTypes.ENUM)
-    ) {
-      const enumType = arraySubtype || attribute.type;
-      const values = enumType.options.values;
-
-      if (Array.isArray(values) && values.length > 0) {
-        type = `ENUM(${values.map(value => this.escape(value)).join(', ')})`;
-
-        if (attribute.type instanceof DataTypes.ARRAY) {
-          type += '[]';
-        }
-      } else {
-        throw new Error("Values for ENUM haven't been defined.");
-      }
-    }
-
-    if (!type) {
-      type = attribute.type;
-    }
-
-    let sql = type.toString();
-
-    if (attribute.allowNull === false) {
-      sql += ' NOT NULL';
-    }
-
-    if (attribute.autoIncrement) {
-      if (attribute.autoIncrementIdentity) {
-        sql += ' GENERATED BY DEFAULT AS IDENTITY';
-      } else {
-        sql += ' SERIAL';
-      }
-    }
-
-    if (defaultValueSchemable(attribute.defaultValue, this.dialect)) {
-      sql += ` DEFAULT ${this.escape(attribute.defaultValue, { type: attribute.type })}`;
-    }
-
-    if (attribute.unique === true) {
-      sql += ' UNIQUE';
-    }
-
-    if (attribute.primaryKey) {
-      sql += ' PRIMARY KEY';
-    }
-
-    if (attribute.references) {
-      // an unqualified reference target lives in the same schema as the table that references it
-      const schema =
-        options?.tableOrModel &&
-        (!attribute.references.table || typeof attribute.references.table === 'string')
-          ? this.extractTableDetails(options.tableOrModel).schema
-          : undefined;
-
-      const referencesTable = this.extractTableDetails(attribute.references.table, { schema });
-
-      let referencesKey;
-
-      if (!options?.withoutForeignKeyConstraints) {
-        if (attribute.references.key) {
-          referencesKey = this.quoteIdentifiers(attribute.references.key);
-        } else {
-          referencesKey = this.quoteIdentifier('id');
-        }
-
-        sql += ` REFERENCES ${this.quoteTable(referencesTable)} (${referencesKey})`;
-
-        if (attribute.onDelete) {
-          sql += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
-        }
-
-        if (attribute.onUpdate) {
-          sql += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
-        }
-
-        if (attribute.references.deferrable) {
-          sql += ` ${this.#internals.getDeferrableConstraintSnippet(attribute.references.deferrable)}`;
-        }
-      }
-    }
-
-    if (attribute.comment && typeof attribute.comment === 'string') {
-      if (options?.context === 'addColumn' || options?.context === 'changeColumn') {
-        const quotedAttr = this.quoteIdentifier(attribute.field);
-        const escapedCommentText = this.escape(attribute.comment);
-        sql += `; COMMENT ON COLUMN ${this.quoteTable(options.tableOrModel)}.${quotedAttr} IS ${escapedCommentText}`;
-      } else {
-        // for createTable event which does it's own parsing
-        // TODO: centralize creation of comment statements here
-        sql += ` COMMENT ${attribute.comment}`;
-      }
-    }
-
-    return sql;
-  }
-
-  attributesToSQL(attributes, options) {
-    const result = {};
-
-    for (const key of Object.keys(attributes)) {
-      const rawAttribute = attributes[key];
-      const attribute = isPlainObject(rawAttribute) ? { ...rawAttribute } : { type: rawAttribute };
-      const columnName = attribute.field || attribute.columnName || key;
-
-      attribute.field = columnName;
-
-      result[columnName] = this.attributeToSQL(attribute, options);
-    }
-
-    return result;
   }
 
   createTrigger(
