+++ packages/sqlite3/src/query-generator.js
@@ -1,10 +1,7 @@
 'use strict';
 
 import { ParameterStyle } from '@sequelize/core';
-import {
-  attributeTypeToSql,
-  normalizeDataType,
-} from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
+import { normalizeDataType } from '@sequelize/core/_non-semver-use-at-your-own-risk_/abstract-dialect/data-types-utils.js';
 import {
   ADD_COLUMN_QUERY_SUPPORTABLE_OPTIONS,
   CREATE_TABLE_QUERY_SUPPORTABLE_OPTIONS,
@@ -12,12 +9,10 @@ import {
 import { rejectInvalidOptions } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/check.js';
 import { removeNullishValuesFromHash } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/format.js';
 import { EMPTY_SET } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/object.js';
-import { defaultValueSchemable } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/query-builder-utils.js';
 import { createBindParamGenerator } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/sql.js';
 import { pojo } from '@sequelize/utils';
 import defaults from 'lodash/defaults';
 import each from 'lodash/each';
-import isPlainObject from 'lodash/isPlainObject';
 import { SqliteQueryGeneratorTypeScript } from './query-generator-typescript.internal.js';
 
 export class SqliteQueryGenerator extends SqliteQueryGeneratorTypeScript {
@@ -123,7 +118,7 @@ export class SqliteQueryGenerator extends SqliteQueryGeneratorTypeScript {
       type: normalizeDataType(dataType.type, this.dialect),
     };
 
-    const definition = this.attributeToSQL(dataType, {
+    const definition = this.attributeToSql(dataType, {
       context: 'addColumn',
       tableOrModel: table,
     });
@@ -199,76 +194,6 @@ export class SqliteQueryGenerator extends SqliteQueryGeneratorTypeScript {
     const result = { query };
     if (parameterStyle === ParameterStyle.BIND) {
       result.bind = bind;
-    }
-
-    return result;
-  }
-
-  attributeToSQL(attribute, options) {
-    if (!isPlainObject(attribute)) {
-      attribute = { type: attribute };
-    }
-
-    let sql = attributeTypeToSql(attribute.type);
-
-    if (attribute.allowNull === false) {
-      sql += ' NOT NULL';
-    }
-
-    if (defaultValueSchemable(attribute.defaultValue, this.dialect)) {
-      // TODO thoroughly check that DataTypes.NOW will properly
-      // get populated on all databases as DEFAULT value
-      // i.e. mysql requires: DEFAULT CURRENT_TIMESTAMP
-      sql += ` DEFAULT ${this.escape(attribute.defaultValue, { ...options, type: attribute.type })}`;
-    }
-
-    if (attribute.unique === true) {
-      sql += ' UNIQUE';
-    }
-
-    if (attribute.primaryKey) {
-      sql += ' PRIMARY KEY';
-
-      if (attribute.autoIncrement) {
-        sql += ' AUTOINCREMENT';
-      }
-    }
-
-    if (!options?.withoutForeignKeyConstraints && attribute.references) {
-      const referencesTable = this.quoteTable(attribute.references.table);
-
-      let referencesKey;
-      if (attribute.references.key) {
-        referencesKey = this.quoteIdentifier(attribute.references.key);
-      } else {
-        referencesKey = this.quoteIdentifier('id');
-      }
-
-      sql += ` REFERENCES ${referencesTable} (${referencesKey})`;
-
-      if (attribute.onDelete) {
-        sql += ` ON DELETE ${attribute.onDelete.toUpperCase()}`;
-      }
-
-      if (attribute.onUpdate) {
-        sql += ` ON UPDATE ${attribute.onUpdate.toUpperCase()}`;
-      }
-    }
-
-    return sql;
-  }
-
-  attributesToSQL(attributes, options) {
-    const result = {};
-
-    for (const key of Object.keys(attributes)) {
-      const rawAttribute = attributes[key];
-      const attribute = isPlainObject(rawAttribute) ? { ...rawAttribute } : { type: rawAttribute };
-      const columnName = attribute.field || attribute.columnName || key;
-
-      attribute.field = columnName;
-
-      result[columnName] = this.attributeToSQL(attribute, options);
     }
 
     return result;
